Adding CAN reading and state control,  sensor reading
Tested the software for writing and reading the CANbus, implemented what I found worked here.

NOT TESTED

Began to work on the state control based on a message recieved from the master board.
Began to work on reading the sensor values in the collect state. Need to add the write to the CANBus